refactor pcb trace element building logic into a single buildPcbTraceElements fn
